fix: resolve Intel compiler errors and warnings
- Add bind(c) to dgemm_blas and zgemm_blas interfaces in sparseir.f90
  to match bind(c) attributes in sparseir_ext.F90 (fixes error #8128)
- Convert status_c to INT(status_c) in all errore calls to match
  INTEGER type expected by errore subroutine (fixes warning #6075)
- Update all code generation scripts to generate INT(status_c) in errore calls
- Regenerate all implementation files (*_impl.inc) with corrected errore calls

+++ fortran/script/generate_dlr2ir.py
@@ -86,7 +86,7 @@ def generate_dlr2ir_function(ndim, input_type, output_type):
   status_c = {c_function}(obj%dlr_f_ptr, obj%backend_ptr, SPIR_ORDER_COLUMN_MAJOR, &
     ndim_c, c_loc(input_dims_c), target_dim_c, c_loc(arr), c_loc(res))
   IF (status_c /= 0) THEN
-    CALL errore('dlr2ir_{func_suffix}_{ndim}d', 'Error converting DLR to IR', status_c)
+    CALL errore('dlr2ir_{func_suffix}_{ndim}d', 'Error converting DLR to IR', INT(status_c))
   ENDIF
 END SUBROUTINE"""
